fix: signet nodes for quick open channel
